Refractor some typescript, added the interfaces, enum and other stuff in a types folder.
