+++ server/types/db.types.ts
@@ -1,0 +1,10 @@
+import { Document } from "mongoose";
+
+export interface IHighscore extends Document {
+  name: string;
+  time: number;
+  guessCount: number;
+  wordLength: number;
+  uniqueLetters: boolean;
+  date: Date;
+}
