+++ server/types/feedback.types.ts
@@ -1,0 +1,11 @@
+export enum LetterResult {
+  CORRECT = "correct",
+  MISPLACED = "misplaced",
+  INCORRECT = "incorrect",
+  PENDING = "pending",
+}
+
+export interface LetterFeedback {
+  letter: string;
+  result: LetterResult;
+}
