+++ server/utils/highscores.ts
@@ -1,5 +1,5 @@
-import Highscore, { IHighscore } from "../models/Highscore.js";
-
+import Highscore from "../models/Highscore.js";
+import { IHighscore } from "../types/db.types.js";
 //The function takes a Partial<IHighscore> object, which means all fields in iHighscore are "free of choice"
 
 export async function addHighscore(
@@ -18,17 +18,6 @@ export async function addHighscore(
     throw error;
   }
 }
-
-// export async function getHighscores(): Promise<IHighscore[]> {
-//   try {
-//     //Gets all highscores from the database and sort them by time.
-//     const highscores = await Highscore.find().sort({ time: 1 });
-//     return highscores;
-//   } catch (error) {
-//     console.error("Error getting highscores:", error);
-//     throw error;
-//   }
-// }
 
 export async function getFilteredHighscores(
   wordLength: number | null = null, //wordlength can be either a nymber or null, default is null.
