+++ server/utils/wordleFeedback.ts
@@ -1,14 +1,4 @@
-export enum LetterResult {
-  CORRECT = "correct",
-  MISPLACED = "misplaced",
-  INCORRECT = "incorrect",
-  PENDING = "pending",
-}
-
-export interface LetterFeedback {
-  letter: string;
-  result: LetterResult;
-}
+import { LetterResult, LetterFeedback } from "../types/feedback.types.js";
 
 export default function wordleFeedback(
   guess: string,
